FerriteNeumann

Documentation for FerriteNeumann.

FerriteNeumann.Neumann

FerriteNeumann.Neumann — Type

Neumann(field_name::Symbol, fv::FaceValues, faceset::Set{FaceIndex}, f, cellset=nothing)

Define a Neumann contribution according to 
```math
\int_{\Gamma} b \ \delta u \ \mathrm{d}\Gamma

\int_{\Gamma} \boldsymbol{b} \cdot \boldsymbol{\delta u} \ \mathrm{d}\Gamma
```
for scalar or vector fields, respectively. Here ``b`` or 
``\boldsymbol{b}`` is the prescribed Neumann value, 
defined by the function `f` with signature 
``` 
f(x::Vec, time, n::Vec)
```
where `x` is the spatial position of the current quadrature point, `time` is the 
current time, and `n` is the face normal vector. `field_name` describes the 
field on which the boundary condition should act, and `fv` describes the 
interpolation and integration rules. `faceset` describes which faces the BC is 
applied to, and if `cellset!=nothing`, only the cells in `faceset` that are also 
in `cellset` are included (important when using mixed grids with different cells)